refactor(typing): reduce aws_lambda_powertools.shared.types usage (#4896)
* refactor(typing): reduce aws_lambda_powertools.shared.types usage

As discussed in #4607. This simplifies linting and refactoring so we can
introduce

    from __future__ import annotations

to all files, which is the plan as the next step.

* Docmentation fix + small changes

---------
