+++ aws_lambda_powertools/event_handler/api_gateway.py
@@ -17,6 +17,7 @@ from typing import (
     Dict,
     Generic,
     List,
+    Literal,
     Mapping,
     Match,
     Optional,
@@ -47,7 +48,6 @@ from aws_lambda_powertools.event_handler.util import _FrozenDict, extract_origin
 from aws_lambda_powertools.shared.cookies import Cookie
 from aws_lambda_powertools.shared.functions import powertools_dev_is_set
 from aws_lambda_powertools.shared.json_encoder import Encoder
-from aws_lambda_powertools.shared.types import Literal
 from aws_lambda_powertools.utilities.data_classes import (
     ALBEvent,
     APIGatewayProxyEvent,
