+++ aws_lambda_powertools/event_handler/middlewares/base.py
@@ -1,9 +1,8 @@
 from abc import ABC, abstractmethod
-from typing import Generic
+from typing import Generic, Protocol
 
 from aws_lambda_powertools.event_handler.api_gateway import Response
 from aws_lambda_powertools.event_handler.types import EventHandlerInstance
-from aws_lambda_powertools.shared.types import Protocol
 
 
 class NextMiddleware(Protocol):
