+++ aws_lambda_powertools/utilities/feature_flags/appconfig.py
@@ -4,16 +4,15 @@ from typing import Any, Dict, Optional, Union, cast
 
 from botocore.config import Config
 
+from aws_lambda_powertools.logging import Logger
 from aws_lambda_powertools.utilities import jmespath_utils
+from aws_lambda_powertools.utilities.feature_flags.base import StoreProvider
+from aws_lambda_powertools.utilities.feature_flags.exceptions import ConfigurationStoreError, StoreClientError
 from aws_lambda_powertools.utilities.parameters import (
     AppConfigProvider,
     GetParameterError,
     TransformParameterError,
 )
-
-from ... import Logger
-from .base import StoreProvider
-from .exceptions import ConfigurationStoreError, StoreClientError
 
 
 class AppConfigStore(StoreProvider):
