+++ aws_lambda_powertools/utilities/parser/types.py
@@ -1,10 +1,8 @@
 """Generics and other shared types used across parser"""
 
-from typing import Any, Dict, Type, TypeVar, Union
+from typing import Any, Dict, Literal, Type, TypeVar, Union
 
 from pydantic import BaseModel, Json
-
-from aws_lambda_powertools.shared.types import Literal
 
 Model = TypeVar("Model", bound=BaseModel)
 EnvelopeModel = TypeVar("EnvelopeModel")
