+++ aws_lambda_powertools/utilities/streaming/s3_object.py
@@ -7,6 +7,7 @@ from typing import (
     Any,
     Iterable,
     List,
+    Literal,
     Optional,
     Sequence,
     TypeVar,
@@ -15,7 +16,6 @@ from typing import (
     overload,
 )
 
-from aws_lambda_powertools.shared.types import Literal
 from aws_lambda_powertools.utilities.streaming._s3_seekable_io import _S3SeekableIO
 from aws_lambda_powertools.utilities.streaming.transformations import (
     CsvTransform,
