+++ examples/event_handler_bedrock_agents/src/assert_bedrock_agent_response_module.py
@@ -1,9 +1,10 @@
 import time
 
+from typing_extensions import Annotated
+
 from aws_lambda_powertools import Logger, Tracer
 from aws_lambda_powertools.event_handler import BedrockAgentResolver
 from aws_lambda_powertools.event_handler.openapi.params import Body
-from aws_lambda_powertools.shared.types import Annotated
 from aws_lambda_powertools.utilities.typing import LambdaContext
 
 tracer = Tracer()
