+++ examples/event_handler_graphql/src/async_resolvers.py
@@ -1,12 +1,11 @@
 import asyncio
-from typing import List
+from typing import List, TypedDict
 
 import aiohttp
 
 from aws_lambda_powertools import Logger, Tracer
 from aws_lambda_powertools.event_handler import AppSyncResolver
 from aws_lambda_powertools.logging import correlation_paths
-from aws_lambda_powertools.shared.types import TypedDict
 from aws_lambda_powertools.tracing import aiohttp_trace_config
 from aws_lambda_powertools.utilities.typing import LambdaContext
