+++ examples/event_handler_graphql/src/getting_started_graphql_api_resolver.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import List, TypedDict
 
 import requests
 from requests import Response
@@ -6,7 +6,6 @@ from requests import Response
 from aws_lambda_powertools import Logger, Tracer
 from aws_lambda_powertools.event_handler import AppSyncResolver
 from aws_lambda_powertools.logging import correlation_paths
-from aws_lambda_powertools.shared.types import TypedDict
 from aws_lambda_powertools.utilities.data_classes.appsync import scalar_types_utils
 from aws_lambda_powertools.utilities.typing import LambdaContext
