+++ examples/event_handler_rest/src/working_with_multi_query_values.py
@@ -1,9 +1,10 @@
 from enum import Enum
 from typing import List
 
+from typing_extensions import Annotated
+
 from aws_lambda_powertools.event_handler import APIGatewayRestResolver
 from aws_lambda_powertools.event_handler.openapi.params import Query
-from aws_lambda_powertools.shared.types import Annotated
 from aws_lambda_powertools.utilities.typing import LambdaContext
 
 app = APIGatewayRestResolver(enable_validation=True)
