+++ tests/performance/parser/test_parser_performance.py
@@ -1,11 +1,11 @@
 import time
 from contextlib import contextmanager
-from typing import Generator
+from typing import Generator, Literal, Union
 
 import pytest
 from pydantic import BaseModel, Field
+from typing_extensions import Annotated
 
-from aws_lambda_powertools.shared.types import Annotated, Literal, Union
 from aws_lambda_powertools.utilities.parser import parse
 
 # adjusted for slower machines in CI too
